+++ NEWS.md
@@ -6,6 +6,10 @@
 * Hash the node names and quote the label names of `tar_mermaid()` graphs to avoid JavaScript keywords.
 * Remove superfluous line breaks in the node labels of graph visuals.
 
+# Enhancements
+
+* Load packages when loading data for downstream targets in the pipeline (#713).
+
 # targets 0.11.0
 
 ## Bug fixes